LayoutTests/imported/w3c/web-platform-tests/IndexedDB/storage-buckets.https.any.serviceworker-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: NetworkError: Load failed

Harness Error (FAIL), message = Unhandled rejection: NetworkError: Load failed
